requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0
smtplib-ssl==1.0.0
email-validator==2.1.0
asyncio==3.4.3
aiofiles==23.2.1